restore New Issue template
Gitlab appears to have introduced a new UI/system that encompasses issue and
Merge Request templating. The docs¹ say that this does not take priority over
existing Gitlab settings:

  Create an issue template named `Default.md` (case-insensitive) and save it in
  `.gitlab/issue_templates/`. The `Default.md` template does not take priority
  over the default template set in project settings.

However this appears to be false. https://gitlab.com/graphviz/graphviz still has
a configured default template, yet now shows a blank text field when creating a
new issue. Lets see if explicitly creating the content of the existing template
in the new way restores it…

This change includes the content of the existing template, with the British
English “behaviour” converted to US English “behavior.”

¹ https://docs.gitlab.com/user/project/description_templates/#set-a-default-template-for-merge-requests-and-issues

Gitlab: closes #2730