CI: Add known-broken Fedora 43 configs
Fedora 43 was released two weeks ago on 2025-10-30. It brings with it
GDK Pixbuf 2.44, which has a new Rust-based image loader named Glycin2
which is **incompatible** with the graphviz `gdk` plugin. (See issue
graphviz/graphviz#2732.) The plugin will silently produce empty files
for any image exports attempted using it.

Since a test was added for this condition in graphviz/graphviz!4555,
it is expected that these Fedora 43 builds will _fail_ until the
gdk plugin is either fixed or disabled in Fedora 43 builds.